add controller

+++ src/main/java/com/example/video/service/impl/JpaKomentarService.java
@@ -6,6 +6,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import com.example.video.model.Komentar;
+import com.example.video.model.Video;
 import com.example.video.repository.KomentarRepository;
 import com.example.video.service.KomentarService;
 
@@ -28,8 +29,17 @@ public class JpaKomentarService implements KomentarService {
 	}
 
 	@Override
-	public void delete(Long id) {
-		 kometarRepository.deleteById(id);
+	public Komentar delete(Long id) {
+java.util.Optional<Komentar> por=  kometarRepository.findById(id);
+		
+		if (por.isPresent()){
+			kometarRepository.deleteById(id);
+		//porRep.delete(por.get());
+			
+			
+			return por.get();
+		}
+		return null;
 		
 	}
